refactor: rename mpc package to trajopt, also tuned ltv
