+++ TeamCode/src/main/java/sigmacorns/opmode/auto/Auto12Ball.kt
@@ -5,7 +5,7 @@ import sigmacorns.Robot
 import sigmacorns.State
 import sigmacorns.constants.drivetrainParameters
 import sigmacorns.control.ltv.LTVClient
-import sigmacorns.control.mpc.TrajoptLoader
+import sigmacorns.control.trajopt.TrajoptLoader
 import sigmacorns.io.JoltSimIO
 import sigmacorns.io.SIM_UPDATE_TIME
 import sigmacorns.io.BallColor
@@ -47,6 +47,7 @@ class Auto12Ball : SigmaOpMode() {
         val initialSample = traj.getInitialSample()
             ?: throw IllegalStateException("Trajectory has no samples")
 
+        io.setPosition(initialSample.pos)
         // Preload 3 balls in sim
         val joltSim = io as? JoltSimIO
         repeat(3) { joltSim?.heldBalls?.add(BallColor.GREEN) }
@@ -54,6 +55,7 @@ class Auto12Ball : SigmaOpMode() {
         val robot = Robot(io, blue = true)
         robot.init(initialSample.pos, apriltagTracking = !SIM)
         robot.aimTurret = true
+        robot.aimFlywheel = false
 
         // Sim X axis is negated relative to real field — flip goal for aiming
         if (SIM) {
@@ -76,29 +78,28 @@ class Auto12Ball : SigmaOpMode() {
 
             // Phase 1: Spin up flywheel then shoot preloads via TRANSFERRING
             robot.aimFlywheel = false
-            robot.shooter.flywheelTarget = FLYWHEEL_TARGET
-            robot.intakeTransfer.state = IntakeTransfer.State.IDLE
-            val preShootStart = io.time()
-            while (opModeIsActive()) {
-                state.update(io)
-                val t = (io.time() - preShootStart).toDouble(DurationUnit.SECONDS)
-
-                // Spin up for 1s, then transfer for 1.5s to shoot all 3
-                if (t >= 1.0) {
-                    robot.intakeTransfer.state = IntakeTransfer.State.TRANSFERRING
-                }
-
-                robot.update()
-                io.update()
-                if (SIM) sleep(SIM_UPDATE_TIME.inWholeMilliseconds)
-                if (t >= 1.0 + shootDuration) break
-            }
+//            robot.shooter.flywheelTarget = FLYWHEEL_TARGET
+//            robot.intakeTransfer.state = IntakeTransfer.State.IDLE
+//            val preShootStart = io.time()
+//            while (opModeIsActive()) {
+//                state.update(io)
+//                val t = (io.time() - preShootStart).toDouble(DurationUnit.SECONDS)
+//
+//                // Spin up for 1s, then transfer for 1.5s to shoot all 3
+//                if (t >= 1.0) {
+////                    robot.intakeTransfer.state = IntakeTransfer.State.TRANSFERRING
+//                }
+//
+//                robot.update()
+//                io.update()
+//                if (SIM) sleep(SIM_UPDATE_TIME.inWholeMilliseconds)
+//                if (t >= 1.0 + shootDuration) break
+//            }
 
             // Phase 2: Follow trajectory with intake/shoot cycles
-            robot.intakeTransfer.state = IntakeTransfer.State.IDLE
-            robot.shooter.flywheelTarget = 0.0
+//            robot.intakeTransfer.state = IntakeTransfer.State.IDLE
+//            robot.shooter.flywheelTarget = 0.0
 
-            io.setPosition(initialSample.pos)
             val startTime = io.time()
 
             while (opModeIsActive()) {
@@ -130,15 +131,15 @@ class Auto12Ball : SigmaOpMode() {
                     elapsedSeconds >= shootTime - spinUpLeadTime && elapsedSeconds < shootTime
                 }
                 robot.aimFlywheel = false
-                robot.shooter.flywheelTarget = if (shouldSpinUp) FLYWHEEL_TARGET else 0.0
+//                robot.shooter.flywheelTarget = if (shouldSpinUp) FLYWHEEL_TARGET else 0.0
 
                 // State machine: intake during intake zones, transfer during shoot windows
                 when {
                     isShooting -> {
-                        robot.intakeTransfer.state = IntakeTransfer.State.TRANSFERRING
+//                        robot.intakeTransfer.state = IntakeTransfer.State.TRANSFERRING
                     }
                     traj.isIntakeZone(elapsedSeconds) -> {
-                        robot.intakeCoordinator.startIntake()
+//                        robot.intakeCoordinator.startIntake()
                     }
                     else -> {
                         if (robot.intakeTransfer.state == IntakeTransfer.State.INTAKING ||
