+++ TeamCode/src/test/kotlin/sigmacorns/test/LTVSolverTimingTest.kt
@@ -6,7 +6,7 @@ import sigmacorns.State
 import sigmacorns.constants.drivetrainParameters
 import sigmacorns.control.ltv.LTVClient
 import sigmacorns.control.ltv.QpSolverType
-import sigmacorns.control.mpc.TrajoptLoader
+import sigmacorns.control.trajopt.TrajoptLoader
 import sigmacorns.io.RerunLogging
 import sigmacorns.io.SimIO
 import sigmacorns.opmode.SigmaOpMode
